improv: use union of strokes when comparing them
Improved the strokes comparison algorithm by calculating the
symmetric hausdorff distance only between the set of points
represented by the union of the strokes.

The natural logs of the width:height ratio of the strokes is
still used for comparison as well, but the log of the ratio
of the union of the strokes is used as well.

This ensures the strokes are not only similar, but also positioned
similarly relative to each other.