change: slightly move mouse over widget popup menu
This improves user experience by making clicking the first option
faster.

We already make it so the first option in the popup menu appears under
the mouse, but the user still had to move the mouse slightly for the
option to be highlighted and for the user to be able to click it.

Now the user only has to click when the menu appears because the slight
movement causes the first option to be selected.