dependency: use shapely instead of scipy
The reason: scipy has an operation to calculate the Hausdorff distance
that we use to match the drawings, but has no available operation to
scale geometry, which we need for a feature we are about to implement.

That feature makes it so the widgets take the size of the drawing.
For us to be able to do implement that we need to be able to scale the
drawings made by users on canvas so they can be compared.

Shapely, on the other hand, has operations for both needs (Hausdorff
distance and geometry resizing/scaling). It is more complete for our
purpose, which is why it is replacing scipy.