misc: use shapely to add and improve functionality
Improved functionality: using shapely.affinity.translate to offset
strokes (so the first point in the first stroke is always at (0, 0))
instead of doing it manually.

Added functionality: using shapely.affinity.scale to scale the strokes
drawn by the user on the canvas so they match the size of the strokes
set for each potentially-matching widget.